+++ R/gggenes-package.R
@@ -2,6 +2,5 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom lifecycle deprecate_warn
 ## usethis namespace: end
 NULL